JUnit test and printing tree feature are added
JUnit test is added:
For Parser class methods test, JUnit test class is added.
This can be used for each method of functionality.
Access level modifier of every method of Parser class
is changed to public for the test.
This should be back after the test is done.

Printing Node Tree feature is added:
With printTree() method of Node class, you can print node tree.

Assertion bug is fixed:
In java, to use assertion '-ea' option should be added to VM argument.
Assertion should be without NOT for this logic. This is fixed.